Optimize for the situation in which damage reports are coming in rapidly.
Using x11perf, it was possible to get to a point where the scanner queue
started to grow to alarming sizes.  By using a region and dropping any
draws that are going to be redundant, we can prevent that growth.

Essentially, this flows from the fact that relaying data to spice is slower
than X can draw.